% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavinteract-package.R
\docType{package}
\name{lavinteract}
\alias{lavinteract}
\alias{lavinteract-package}
\title{Interaction Testing and Plotting for \code{lavaan} Fitted Objects}
\description{
Tools to test and visualize moderation in structural equation models fitted
with \code{lavaan} using model estimates and their covariance without refitting. 
The framework accepts continuous or categorical moderators.
}
\details{
The function is:
\itemize{
  \item \code{\link{lav_slopes}}: simple slopes and interaction plots from a fitted \code{lavaan} model.
}
}
\section{Note}{

The development of this package was possible only thanks to the constant interaction (sic)
with Dr. Cataldo Giuliano Gemmano
(\href{https://orcid.org/0000-0003-4519-8594}{ORCID: 0000-0003-4519-8594}), and 
particularly because of his sustained feedback and support.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/g-corbelli/lavinteract}
  \item Report bugs at \url{https://github.com/g-corbelli/lavinteract/issues}
}

}
\author{
Giuseppe Corbelli (<giuseppe.corbelli@uniroma1.it>, or <giuseppe.corbelli@uninettunouniversity.net>)
}
\keyword{SEM}
\keyword{interactions}
\keyword{lavaan}
\keyword{moderation}
\keyword{plotting}
\keyword{simple-slopes}
